Review Code

+++ src/main/java/com/example/consolelog/service/CommentService.java
@@ -5,6 +5,7 @@ import com.example.consolelog.dto.responseDto.CommentResponseDto;
 import com.example.consolelog.dto.responseDto.ResponseDto;
 import com.example.consolelog.entity.Board;
 import com.example.consolelog.entity.Comment;
+import com.example.consolelog.entity.Member;
 import com.example.consolelog.repository.BoardRepository;
 import com.example.consolelog.repository.CommentRepository;
 import com.example.consolelog.repository.MemberRepository;
@@ -16,21 +17,15 @@ import javax.transaction.Transactional;
 @Service
 @RequiredArgsConstructor
 public class CommentService {
-
-    private final MemberRepository memberRepository;
     private final BoardRepository boardRepository;
     private final CommentRepository commentRepository;
 
     // 댓글 생성
-    public ResponseDto<?> createComment(Long boardId, CommentRequestDto commentRequestDto, MemberDetailsImpl memberDetails) {
+    public ResponseDto<?> createComment(Long boardId, CommentRequestDto commentRequestDto, Member member) {
 
-        memberRepository.findByName(memberDetails.getUsername()).orElseThrow(()
-                -> new NullPointerException("해당 사용자의 정보가 존재하지 않습니다."));
+        Board board = boardRepository.findById(boardId).orElseThrow(() -> new NullPointerException("해당 게시물이 존재하지 않습니다."));
 
-        Board board = boardRepository.findById(boardId).orElseThrow(()
-                -> new NullPointerException("해당 게시물이 존재하지 않습니다."));
-
-        Comment comment = new Comment(commentRequestDto, memberDetails.getMember(), board);
+        Comment comment = new Comment(commentRequestDto, member, board);
 
         commentRepository.save(comment);
 
@@ -39,13 +34,13 @@ public class CommentService {
 
     // 댓글 수정
     @Transactional
-    public ResponseDto<?> updateComment(Long commentId, CommentRequestDto commentRequestDto, MemberDetailsImpl memberDetails) {
+    public ResponseDto<?> updateComment(Long commentId, CommentRequestDto commentRequestDto, Member member) {
 
-        memberRepository.findByName(memberDetails.getUsername()).orElseThrow(()
-                -> new NullPointerException("해당 사용자의 정보가 존재하지 않습니다."));
 
-        Comment comment = commentRepository.findById(commentId).orElseThrow(()
-                -> new NullPointerException("해당 댓글이 존재하지 않습니다."));
+        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new NullPointerException("해당 댓글이 존재하지 않습니다."));
+
+        if (validateMember(member, comment))
+            throw new IllegalArgumentException("댓글 작성자와 현재 사용자가 일치하지 않습니다.");
 
         comment.update(commentRequestDto);
 
@@ -54,16 +49,20 @@ public class CommentService {
 
     // 댓글 삭제
     @Transactional
-    public ResponseDto<?> deleteComment(Long commentId, MemberDetailsImpl memberDetails) {
+    public ResponseDto<?> deleteComment(Long commentId, Member member) {
 
-        memberRepository.findByName(memberDetails.getUsername()).orElseThrow(()
-                -> new NullPointerException("해당 사용자의 정보가 존재하지 않습니다."));
+        Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new NullPointerException("해당 댓글이 존재하지 않습니다."));
 
-        Comment comment = commentRepository.findById(commentId).orElseThrow(()
-                -> new NullPointerException("해당 댓글이 존재하지 않습니다."));
+        if (validateMember(member, comment))
+            throw new IllegalArgumentException("댓글 작성자와 현재 사용자가 일치하지 않습니다.");
 
         commentRepository.delete(comment);
 
         return ResponseDto.success("댓글 삭제 완료");
+    }
+
+    public boolean validateMember(Member member, Comment comment) {
+
+        return !member.getName().equals(comment.getMember().getName());
     }
 }
